requirements.txt
django-filter
appdirs==1.4.4
asgiref==3.2.3
autopep8==1.5.4
boto3==1.12.42
botocore==1.15.42
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
defusedxml==0.6.0
distlib==0.3.1
dj-database-url==0.5.0
Django==3.0.1
django-allauth==0.41.0
django-countries==6.0
django-crispy-forms==1.8.1
django-storages==1.9.1
dnspython==2.0.0
docutils==0.15.2
filelock==3.0.12
Flask==1.1.2
Flask-PyMongo==2.3.0
gunicorn==20.0.4
idna==2.8
itsdangerous==1.1.0
Jinja2==2.11.2
jmespath==0.9.5
MarkupSafe==1.1.1
migrate==0.3.8
oauthlib==3.1.0
Pillow==7.0.0
psycopg2-binary==2.8.5
pycodestyle==2.6.0
pymongo==3.11.0
PyMySQL==0.10.1
python-dateutil==2.8.1
python3-openid==3.1.0
pytz==2019.3
requests==2.22.0
requests-oauthlib==1.3.0
s3transfer==0.3.3
six==1.15.0
sqlparse==0.3.0
stripe==2.42.0
toml==0.10.1
urllib3==1.25.7
virtualenv==20.0.31
Werkzeug==1.0.1
